Add --monitor check and removed get_monitors.py with improved replacement code
--moniter argument checks now fails if --wallpaper or --random is not provided.\nRemoved and replaced get_monitors.py with authors suggestion. Now sets MONITORS_OPTIONS with one line in options.py instead of using a whole other module file.
